chore: Break Gitea e2e tests down
Break down the monolithic Gitea E2E test file (42 tests) into
8 focused files grouped by concern: remote tasks, concurrency,
YAML errors, error snippets, push events, CLI, event filtering,
and provenance. No test behavior changes.

Fixes #2828
